Move Seed Integration jobs into seed package
